+++ pyviews/core/__init__.py
@@ -6,5 +6,4 @@ from .compilation import Expression, ObjectNode, CompilationError
 from .node import Node, InstanceNode, Modifier, Property
 from .observable import Observable, ObservableEntity, InheritedDict
 from .reflection import import_path
-from .rendering import create_node, render
 from .xml import XmlAttr, XmlNode